pyproject.toml
[tool.poetry]
name = "electronicssales"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
flake8 = "^7.2.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''

[tool.isort]
line_length = 119
